Set windows on-screen to the correct size
This ensures that windows on-screen are set to the correct size as
specified by the geometry (if provided by the client) or the size of the
toplevel buffer (otherwise).

Additionally, the arguments passed to the agent are now interpreted as a
program and arguments to be run in a subprocess.