Working keyboard input
As of this commit keyboard input is almost fully functional.  There is a
major limitation, though, in that newly focused events don’t receive a
list of currently-pressed keys.  This is a bug and will be fixed.

This commit is made using Vim in the “foot” terminal emulator in this
compositor.  Our compositor is finally able to self-host!