Release keys in response to MSG_KEYMAP_NOTIFY
If a window loses keyboard focus while a key is pressed, the GUI agent
will not be notified that the key has been released.  The GUI daemon
works around this problem by sending MSG_KEYMAP_NOTIFY events whenever a
window gains keyboard or pointer focus.  When these are sent, release
any keys that need to be released.