Avoid a potential use-after-free on display destruction
The data source is owned by the wlr_seat and must not be destroyed by
any other code.  Destroying the wl_display will destroy the wlr_seat,
which will in turn destroy the qubes_output_source.

qubes_clipboard_writer still needs a listener for display destruction,
since it is owned by the display event loop and the only other way to
free it is for it to write out all of its data.