Initialize writer->fd
Previously this was left at 0, causing file descriptor 0 to be
incorrectly closed instead of the file descriptor used for writing data.
As a result, if the data source was the global clipboard, pasting did
not finish until the client timed out.  Furthermore, the next client to
connect was assigned file descriptor 0.  Therefore, the *next* paste
that had the global clipboard as its source disconnected that client.