Support for reconnecting to the GUI daemon
The GUI daemon can exit for various reasons, such as the user logging
out or an explicit ‘qvm-xkill’ command.  It is possible to restart the
GUI daemon, which will happen when the user logs back in for example.
This requires implementing a state machine to keep track of the current
state of the connection.

The state machine is implemented in Rust, but there is new code on the C
side for reestablishing the current connection state.