Report grant alloc limit on gntalloc failure
This limit is set to an unusably low value (1024) by default.  The X11
GUI agent calls a shell script that rases it, but I am not sure whether
the Wayland agent should do the same.  It is also possible for a script
to raise the limit before running the agent.