Destroy a popup that loses focus
The GUI daemon does not tell the agent about clicks on other qube’s
windows, for obvious security reasons.  Therefore, loss of focus is the
last opportunity the agent has to dismiss a popup before it covers stuff
it really shouldn’t.  This is a deliberate change in behavior from the
X11 GUI agent, but I consider it an improvement.

This reverts commit 480856e65fdc878a902d411430cdcdfeb9ed742f.