Cleanup configure event handling
Configure event handling is known to be buggy.  This is an attempt at
rationalizing it.  The unnecessary QUBES_OUTPUT_NEED_CONFIGURE flag has
been removed, and with it the qubes_output_need_configure function.
Configure events from the GUI daemon are now only ACKd when the client
ACKs the configure event sent by the compositor.