GUI daemon state is authoritative
The GUI daemon’s view of the window state is authoritative; the GUI
agent’s state must be updated from the daemon.  In particular, the GUI
agent MUST NOT update the window size except in response to messages
from the GUI daemon.  If a client requests to resize a window, the
request should be forwarded to the GUI daemon, which will forward it to
the GUI qube’s window manager.  The GUI qube’s window manager will make
the final decision.

This may not completely fix the race condition when a window is rapidly
resized, but it certainly makes the race much harder to trigger.  It
also ensures that the agent’s state is always consistent with the
daemon’s state, and therefore that ‘xdg_surface.configure’ events and
surface drawing reflects the actual state of the window.