Rip out Xwayland support
X11 will be handled via the existing X11 agent and a proxy (included in
the compositor) that multiplexes traffic from different agents.  This
avoids having to work with wlroots's Xwayland support code, allows
writing more code in Rust rather than in C, and should perform better
too.  It also allows multiple X servers to run at the same time.  This
decision may need to be reconsidered if the standalone X server goes
away, but in the meantime it is clearly the correct call.