Fully asynchronous connection establishment
This pulls in fixes to gui-rust that make connection establishment fully
asynchronous.  There is no longer any difference between an initial
connection by the GUI daemon and subsequent reconnections.  In fact, the
agent will happily allow clients to connect even if the GUI daemon has
never connected!  This works out fine because the existing reconnection
code takes care of getting the GUI daemon in sync with the current
agent-side state.