Implement XWayland support
It does not work very well (xterm draws itself at an incorrect
position), but this commit was made with it!  Mouse input does not
work at all.  The problem was that configure events from the GUI daemon
were not properly acknowledged, with the result that the GUI daemon
ignored the drawing commands.  I spent two weeks working on it and gave
up, but then a last-ditch refactor made the bug obvious.