New state tracking
FIXME: write this down!

Fix input event positioning

Input events are at the correct position and must not be adjusted.  The
wlr_xdg_surface already knows its absolute position, so adding its
position to that of input events will result in input events at a
position that is equal to the *sum* of the surface position and the
pointer position.

Properly set window positions

This fixes the position of weston-terminal and of popups in
gnome-terminal.

Make output position handling somewhat reasonable

Output position should be centralized in the output struct.  Passing
boxes around is a recipe for bugs that are nigh-impossible to track down
or fix.  Also remove a bunch of unused code and clean up other code.

Fix input event positioning (again)

The offset to use is stored in the scene_output.

Update mode on initial surface creation

Merely setting a custom mode is not sufficient: the mode defaults to
0x0, but that causes the allocator to refuse the allocation, so the code
that would actually update the mode never runs.

Document how state is tracked

This is going to be invaluable for me to understand what is going on.