Fix input event positioning
Input events are at the correct position and must not be adjusted.  The
wlr_xdg_surface already knows its absolute position, so adding its
position to that of input events will result in input events at a
position that is equal to the *sum* of the surface position and the
pointer position.