Read keyboard layout from QubesDB
This ensures that the keyboard used in the guest matches that on the
host.  The keyboard layout is automatically refreshed whenever the
/keyboard-layout QubesDB entry changes.  If the keyboard layout is not
valid, the agent exits to avoid misinterpreting user keystrokes.
Additionally, changes to the GUI domain XID are now detected and logged.

Do not exit by default if keymap compile fails

If keymap compilation fails after startup, do not exit by default.  Also
massively improve the --help output and improve boolean option parsing.